+++ src/leapfrogai_api/routers/openai/helpers.py
@@ -1,42 +1,14 @@
+"""Helper functions for the OpenAI backend."""
 from typing import BinaryIO, Iterator, Union
 
 import grpc
+from openai.types import CompletionUsage as Usage
 import leapfrogai_sdk as lfai
-from leapfrogai_api.backends.openai.types import (
+from leapfrogai_api.routers.openai.types import (
     ChatCompletionResponse,
     ChatDelta,
     ChatStreamChoice,
-    CompletionChoice,
-    CompletionResponse,
-    Usage,
 )
-
-
-async def recv_completion(
-    stream: grpc.aio.UnaryStreamCall[lfai.CompletionRequest, lfai.CompletionResponse],
-):
-    async for c in stream:
-        yield (
-            "data: "
-            + CompletionResponse(
-                id="foo",
-                object="completion.chunk",
-                created=55,
-                model="mpt-7b-8k-chat",
-                choices=[
-                    CompletionChoice(
-                        index=0,
-                        text=c.choices[0].text,
-                        logprobs=None,
-                        finish_reason="stop",
-                    )
-                ],
-                usage=Usage(prompt_tokens=0, completion_tokens=0, total_tokens=0),
-            ).model_dump_json()
-        )
-        yield "\n\n"
-
-    yield "data: [DONE]"
 
 
 async def recv_chat(
@@ -44,6 +16,7 @@ async def recv_chat(
         lfai.ChatCompletionRequest, lfai.ChatCompletionResponse
     ],
 ):
+    """Generator that yields chat completion responses as Server-Sent Events."""
     async for c in stream:
         yield (
             "data: "
@@ -70,6 +43,7 @@ async def recv_chat(
 
 
 def grpc_chat_role(role: str) -> Union[lfai.ChatRole, None]:
+    """Converts a string to a ChatRole."""
     match role:
         case "user":
             return lfai.ChatRole.USER  # type: ignore
@@ -85,6 +59,7 @@ def grpc_chat_role(role: str) -> Union[lfai.ChatRole, None]:
 
 # read_chunks is a helper method that chunks the bytes of a file (audio file) into a iterator of AudioRequests
 def read_chunks(file: BinaryIO, chunk_size: int) -> Iterator[lfai.AudioRequest]:
+    """Reads a file in chunks and yields AudioRequests."""
     while True:
         chunk = file.read(chunk_size)
         if not chunk:
