+++ tests/pytest/leapfrogai_api/test_api.py
@@ -6,7 +6,7 @@ import time
 import pytest
 from fastapi.testclient import TestClient
 
-import leapfrogai_api.backends.openai.types as lfai_types
+import leapfrogai_api.routers.openai.types as lfai_types
 from leapfrogai_api.main import app
 
 # Set environment variables that the TestClient will use
@@ -22,6 +22,7 @@ LFAI_CONFIG_FILEPATH = os.path.join(LFAI_CONFIG_PATH, LFAI_CONFIG_FILENAME)
 
 
 def test_config_load():
+    """Test that the config is loaded correctly."""
     with TestClient(app) as client:
         response = client.get("/models")
 
@@ -33,6 +34,7 @@ def test_config_load():
 
 
 def test_config_delete(tmp_path):
+    """Test that the config is deleted correctly."""
     # move repeater-test-config.yaml to temp dir so that we can remove it at a later step
     tmp_config_filepath = shutil.copyfile(
         LFAI_CONFIG_FILEPATH, os.path.join(tmp_path, LFAI_CONFIG_FILENAME)
@@ -62,13 +64,13 @@ def test_config_delete(tmp_path):
 
 
 def test_routes():
+    """Test that the expected routes are present."""
     expected_routes = {
         "/docs": ["GET", "HEAD"],
         "/healthz": ["GET"],
         "/models": ["GET"],
-        "/openai/v1/completions": ["POST"],
-        "/openai/v1/chat/completions": ["POST"],
         "/openai/v1/models": ["GET"],
+        "/openai/v1/chat/completions": ["POST"],
         "/openai/v1/embeddings": ["POST"],
         "/openai/v1/audio/transcriptions": ["POST"],
     }
@@ -83,6 +85,7 @@ def test_routes():
 
 
 def test_healthz():
+    """Test the healthz endpoint."""
     with TestClient(app) as client:
         response = client.get("/healthz")
         assert response.status_code == 200
@@ -93,37 +96,8 @@ def test_healthz():
     os.environ.get("LFAI_RUN_REPEATER_TESTS") != "true",
     reason="LFAI_RUN_REPEATER_TESTS envvar was not set to true",
 )
-def test_completion():
-    with TestClient(app) as client:
-        # Send request to client
-        input_text = "This is the completion input text."
-        completion_request = lfai_types.CompletionRequest(
-            model="repeater",
-            prompt=input_text,
-        )
-        response = client.post(
-            "/openai/v1/completions", json=completion_request.model_dump()
-        )
-        assert response.status_code == 200
-
-        # parse through the response
-        response_obj = response.json()
-        assert "choices" in response_obj
-
-        # parse the choices from the response
-        response_choices = response_obj.get("choices")
-        assert len(response_choices) == 1
-        assert "text" in response_choices[0]
-
-        # validate that the repeater repeated
-        assert response_choices[0].get("text") == input_text
-
-
-@pytest.mark.skipif(
-    os.environ.get("LFAI_RUN_REPEATER_TESTS") != "true",
-    reason="LFAI_RUN_REPEATER_TESTS envvar was not set to true",
-)
 def test_embedding():
+    """Test the embedding endpoint."""
     expected_embedding = [0.0 for _ in range(10)]
 
     with TestClient(app) as client:
@@ -146,49 +120,6 @@ def test_embedding():
         data_obj = response_obj.get("data")[0]
         assert "embedding" in data_obj
         assert data_obj.get("embedding") == expected_embedding
-
-
-@pytest.mark.skipif(
-    os.environ.get("LFAI_RUN_REPEATER_TESTS") != "true",
-    reason="LFAI_RUN_REPEATER_TESTS envvar was not set to true",
-)
-def test_stream_completion():
-    with TestClient(app) as client:
-        # Send request to client
-        input_text = "This is the completion input text."
-        stream_completion_request = lfai_types.CompletionRequest(
-            model="repeater",
-            prompt=input_text,
-            stream=True,
-        )
-        response = client.post(
-            "/openai/v1/completions", json=stream_completion_request.model_dump()
-        )
-        assert response.status_code == 200
-        assert (
-            response.headers.get("content-type") == "text/event-stream; charset=utf-8"
-        )
-
-        # parse through the streamed response
-        iter_length = 0
-        iter_lines = response.iter_lines()
-        for line in iter_lines:
-            # skip the empty, and non-data lines
-            if ": " in line:
-                strings = line.split(": ", 1)
-
-                # Process all the data responses that is not the sig_stop signal
-                if strings[0] == "data" and strings[1] != "[DONE]":
-                    stream_response = json.loads(strings[1])
-                    assert "choices" in stream_response
-                    choices = stream_response.get("choices")
-                    assert len(choices) == 1
-                    assert "text" in choices[0]
-                    assert choices[0].get("text") == input_text
-                    iter_length += 1
-
-        # The repeater only response with 5 messages
-        assert iter_length == 5
 
 
 @pytest.mark.skipif(
